POET-95 - Adding and fixing unit tests.
